+++ app/src/main/java/com/herewhite/demo/test/ConvertV5Activity.java
@@ -7,10 +7,12 @@ import android.widget.TextView;
 
 import androidx.appcompat.app.AppCompatActivity;
 
+import com.herewhite.demo.DemoAPI;
 import com.herewhite.demo.R;
 import com.herewhite.sdk.ConverterCallbacks;
-import com.herewhite.sdk.converter.ConvertType;
 import com.herewhite.sdk.converter.ConverterV5;
+import com.herewhite.sdk.converter.ImageFormat;
+import com.herewhite.sdk.converter.ConvertType;
 import com.herewhite.sdk.domain.ConversionInfo;
 import com.herewhite.sdk.domain.ConvertException;
 import com.herewhite.sdk.domain.ConvertedFiles;
@@ -46,10 +48,10 @@ public class ConvertV5Activity extends AppCompatActivity {
                             .setResource(resources[i % resources.length])
                             .setType(randomInt(1) == 1 ? ConvertType.Static : ConvertType.Dynamic)
                             .setPreview(randomInt(1) == 1)
-                            .setOutputFormat(randomInt(1) == 1 ? ConverterV5.OutputFormat.PNG : ConverterV5.OutputFormat.JPEG)
+                            .setOutputFormat(randomInt(1) == 1 ? ImageFormat.PNG : ImageFormat.JPEG)
                             .setTimeout(60_000L)
                             .setPoolInterval(2000)
-                            .setSdkToken("WHITEcGFydG5lcl9pZD1OZ3pwQWNBdlhiemJERW9NY0E0Z0V3RTUwbVZxM0NIbDJYV0Ymc2lnPWNiZWExOTMwNzc1NmQyNmU3N2U3M2Q0NWZjNTZiOGIwMWE2ZjU4NDI6YWRtaW5JZD0yMTYmcm9sZT1hZG1pbiZleHBpcmVfdGltZT0xNTg5ODMzNTQxJmFrPU5nenBBY0F2WGJ6YkRFb01jQTRnRXdFNTBtVnEzQ0hsMlhXRiZjcmVhdGVfdGltZT0xNTU4Mjc2NTg5Jm5vbmNlPTE1NTgyNzY1ODg4NDQwMA")
+                            .setSdkToken(new DemoAPI().getSdkToken())
                             .setCallback(callbacks[i])
                             .build();
                     converter.startConvertTask();